requirements.txt
torch
torchnet
numpy
tqdm
pillow
pickle